Position creator 5 \#release

+++ src/main/java/net/mysticcloud/spigot/core/utils/regions/Region.java
@@ -75,7 +75,7 @@ public class Region {
         this.z2 = z2;
     }
 
-    public int getArea() {
-        return (int) (Math.sqrt(Math.pow(x2 - x1, 2)) * Math.sqrt(Math.pow(y2 - y1, 2)) * Math.sqrt(Math.pow(z2 - z1, 2)));
+    public double getArea() {
+        return (Math.sqrt(Math.pow(x2 - x1, 2)) * Math.sqrt(Math.pow(y2 - y1, 2)) * Math.sqrt(Math.pow(z2 - z1, 2)));
     }
 }
